Refactor tests to app directory
