+++ app/tests/test_pages.py
@@ -1,6 +1,6 @@
 from flask import url_for
 
-from tests.factories import PersonFactory
+from app.tests.factories import PersonFactory
 
 
 def test_index(client):
